pyproject.toml
[tool.poetry]
name = "research"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
speechbrain = "^1.0.2"
torchaudio = "^2.5.1"
soundfile = "^0.12.1"
torch = "^2.5.1"
tqdm = "^4.67.1"
scikit-learn = "^1.6.0"
transformers = {extras = ["torch"], version = "^4.47.1"}
datasets = "^3.2.0"
jiwer = "^3.0.5"
python-dotenv = "^1.0.1"
huggingface-hub = "^0.27.1"
kagglehub = "^0.3.9"
uvicorn = "^0.34.0"
fastapi = "^0.115.8"
matplotlib = "^3.10.1"
python-multipart = "^0.0.20"
pymongo = {extras = ["srv"], version = "^4.11.2"}
motor = "^3.7.0"
ffmpeg = "^1.4"
pydub = "^0.25.1"
gitingest = "^0.1.4"
prettytable = "^3.15.1"
pydantic-settings = "^2.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
